requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
playwright>=1.40.0
streamlit>=1.28.0
httpx>=0.25.0
pillow>=10.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.3.0
streamlit-autorefresh>=1.0.1
aiofiles>=23.0.0
asyncio-throttle>=1.0.0
